Touchboost implementation
Added function pointer to cpufreq.h which will be available only for
touchscreen driver and cpu governors.
Upon governor setting, the pointer gets set to the touchboost function the
governor implements.
When exiting the governor, the pointer is set to NULL of course.

Also touchboost frequency was reduced from maximum to max-500MHz